Remove custom Blazor loadBootResource script
The script block responsible for starting Blazor with a custom `loadBootResource` function has been removed. This function was used to fetch precompressed Brotli (.br) files for faster downloads, particularly because GitHub Pages does not natively support Brotli or gzip for .dll files. The removal of this script means that the application will no longer attempt to load these precompressed resources.