Add GitHub Actions workflow and update README.md
Updated README.md to include a new reference link to a blog post about publishing Blazor WebAssembly projects to GitHub Pages and removed the contact email address. Added a new GitHub Actions workflow file (main.yml) to automate the build, test, and deployment process. The workflow triggers on pushes and pull requests to the main branch and includes a job `deploy-to-github-pages` that sets up the .NET Core SDK, installs the .NET WASM Build Tools, publishes the Blazor project, and commits the `wwwroot` folder to the `gh-pages` branch.