Comment out base href rewrite step in GitHub Actions
The step that rewrites the base href in the GitHub Actions workflow
has been commented out. This step used the
`SteveSandersonMS/ghaction-rewrite-base-href@v1` action to modify the
`base_href` in the `index.html` file located at
`dist/BeyondTheLabel/wwwroot/app/index.html`. The rest of the workflow
remains unchanged, including the steps to publish the .NET Core project
and commit the `wwwroot` folder to GitHub Pages.